fix: babysit round 4, 7 findings (5 coderabbit, 2 deferred file splits)
CodeRabbit inline (5):
- CLAUDE.md:147 -- document MCP invoker boundary regression test
  alongside the BaseTool args_model coverage test.
- CLAUDE.md:181 -- replace hardcoded args-file list with generic
  domain-scoped pattern; add note about the
  _validate_args_model_property_alignment safeguard on DomainToolDef
  that catches drift between args_model and the wire schema at
  registration time.
- src/synthorg/api/ws_payloads/_lifecycle.py -- add ge=0 constraints
  on every duration_seconds field (WsCoordinationPhaseCompletedPayload,
  WsCoordinationFailedPayload.total_duration_seconds,
  WsMeetingCompletedPayload.duration_seconds) plus
  WsBudgetAlertPayload.threshold/current. Negative durations cannot
  cross the websocket boundary.
- src/synthorg/memory/tools/knowledge_architect.py:439 -- swap
  KNOWLEDGE_ARCHITECT_DELETE for the new
  KNOWLEDGE_ARCHITECT_DELETE_FAILED constant on the delete-failure
  warning path so failure events follow the same naming convention as
  KNOWLEDGE_ARCHITECT_SEARCH/READ/WRITE/BROWSE_WIKI_FAILED.
- tests/unit/tools/database/test_args.py:29 -- vendor-neutral
  docstring (SQLite to "Database bind protocols").

Deferred file splits (2, completes round 3 deferred items):
- src/synthorg/api/ws_payloads.py (1077 lines) split into a package:
  __init__.py exports the discriminated WsEventPayload union;
  _lifecycle.py carries Task/Agent/Company/Budget/Message/System/
  Approval/Coordination/Meeting (511 lines); _domain.py carries
  Artifact/Project/MemoryFineTune/Client/Request/Review/Simulation/
  Interrupt+Dissent (399 lines); _base.py owns the shared
  PAYLOAD_CONFIG.
- src/synthorg/meta/mcp/domains/_remaining_args.py (835 lines) split
  into a package: __init__.py re-exports keep the historical import
  surface stable for the four domain consumers (communication.py,
  integrations.py, infrastructure.py, memory.py); _communication.py
  (157 lines), _integrations.py (148), _infrastructure.py (274),
  _memory_finetune.py (173) carry the actual model bodies.
- tests/unit/meta/mcp/domains/test_args_smoke.py -- _models_in helper
  now accepts models defined in any direct submodule of a package so
  the smoke test still walks the full set after the split.

Every file is under the 800-line guideline. All checks green:
ruff (88-col, D205), ruff format, mypy strict on touched modules,
pytest unit suite (25,672 passed in 72.82s).
