+++ CLAUDE.md
@@ -144,7 +144,7 @@ See [docs/reference/configuration-precedence.md](docs/reference/configuration-pr
 - **Args models for tools / handlers (#1611)**: every `BaseTool` subclass and every MCP tool registration declares a typed Pydantic args model. Args models live alongside their consumer in domain-scoped `_args.py` modules and use `ConfigDict(frozen=True, allow_inf_nan=False, extra="forbid")`. Two equivalent realisations of that config exist:
   - **MCP domain args** (`src/synthorg/meta/mcp/domains/_*_args.py`) inherit from `_ArgsBase` (just `BaseModel` + the standard `_ARGS_CONFIG`) and reuse `PaginationFields` / `DestructiveGuardrailFields` mixins from `_common_args.py` for the cross-cutting list / destructive-op shapes.
   - **Tool args** (`src/synthorg/tools/<domain>/_args.py`, `synthorg.memory.tools._args`, `synthorg.memory.self_editing_args`, `synthorg.api.ws_payloads`, `synthorg.a2a.rpc_params`) declare `_ARGS_CONFIG = ConfigDict(...)` at module scope and set `model_config = _ARGS_CONFIG` per class. The pagination / destructive mixins are MCP-specific and don't apply outside the MCP handler surface, so tool args don't reach for them.
-  Both patterns produce the same Pydantic config; pick the one that matches the surrounding domain and don't cross the streams. `BaseTool` subclasses declare `args_model: ClassVar[type[BaseModel] | None] = <Args>`; `MCPToolDef` carries the same `args_model` field. The invoker validates raw args dicts against the model before dispatch; handlers see structurally-validated input and skip manual `common_args` extraction for declared fields. A regression test (`tests/unit/tools/test_args_model_coverage.py`) walks the `BaseTool` subclass tree and fails on any tool missing `args_model` (small allowlist for dynamically-shaped tools).
+  Both patterns produce the same Pydantic config; pick the one that matches the surrounding domain and don't cross the streams. `BaseTool` subclasses declare `args_model: ClassVar[type[BaseModel] | None] = <Args>`; `MCPToolDef` carries the same `args_model` field. The invoker validates raw args dicts against the model before dispatch; handlers see structurally-validated input and skip manual `common_args` extraction for declared fields. Regression tests enforce args_model coverage on both surfaces: `tests/unit/tools/test_args_model_coverage.py` walks the `BaseTool` subclass tree and fails on any tool missing `args_model` (small allowlist for dynamically-shaped tools), and `tests/unit/meta/mcp/test_invoker.py` verifies the MCP invoker boundary semantics (`args_model.model_validate` is called before dispatch, malformed shapes return an `invalid_argument` envelope, handler is never invoked on validation failure).
 - **Async concurrency**: prefer `asyncio.TaskGroup` for fan-out/fan-in parallel operations in new code (e.g. multiple tool invocations, parallel agent calls). Prefer structured concurrency over bare `create_task`. Existing code is being migrated incrementally. When running multiple tasks inside a `TaskGroup` where one task's failure should NOT cancel the others (independent workers, classification detectors, notification sinks), wrap each task body in a small `async def` helper that catches `Exception` and returns a safe default (re-raising only `MemoryError`/`RecursionError`); never let a single worker abort the whole group.
 - **Lifecycle synchronization**: services with async `start()` / `stop()` use a dedicated `self._lifecycle_lock: asyncio.Lock` (separate from any hot-path lock) held across the full body of both methods; timed-out stops must mark the service unrestartable. See [docs/reference/lifecycle-sync.md](docs/reference/lifecycle-sync.md) for the full rule and canonical examples.
 - **Untrusted-content fences at LLM call sites (SEC-1)**: any attacker-controllable string interpolated into an LLM prompt must be wrapped via `wrap_untrusted(tag, content)` from `synthorg.engine.prompt_safety`, and the enclosing system prompt must append `untrusted_content_directive(tags)`. See [docs/reference/sec-prompt-safety.md](docs/reference/sec-prompt-safety.md) for the standard tag set, key reference call sites, and the tool-result injection detector.
@@ -178,7 +178,7 @@ SynthOrg exposes 200+ tools across the 15 domain modules under `src/synthorg/met
 - **`common_args.py`**: argument validators/extractors used by the **legacy** dict-based path: `require_arg`, `require_non_blank`, `get_optional_str`, `require_dict`, `parse_time_window`, `parse_str_sequence`, `coerce_pagination`, `actor_id`, `require_actor_id`, `actor_label`. Tools that declare an `args_model` get Pydantic validation at the invoker boundary (#1611) and skip these helpers.
 - **`common_logging.py`**: the three handler-layer log helpers covering the three log paths every domain handler exercises: `log_handler_argument_invalid` (caught `ArgumentValidationError`), `log_handler_invoke_failed` (any other `Exception` from the service layer, with optional correlation kwargs), and `log_handler_guardrail_violated` (caught `GuardrailViolationError` from a destructive-op precondition). Owns a module-scoped logger keyed at `synthorg.meta.mcp.handlers`, one of the rare carve-outs from the `get_logger(__name__)` rule (see "Logging" above).
 
-**Typed args (#1611 Phase 4)**: every domain registration in `src/synthorg/meta/mcp/domains/<domain>.py` now passes `args_model=<DomainArgs>` to the `read_tool` / `write_tool` / `admin_tool` factories. The `MCPToolInvoker` validates the raw arguments dict against the Pydantic model **before** dispatching to the handler; a malformed shape returns a typed `invalid_argument` envelope without ever invoking the handler. Migrated handlers therefore do NOT need to call `require_arg` / `require_non_blank` / `coerce_pagination` for fields the model already validates. Args models live under `src/synthorg/meta/mcp/domains/_*_args.py` (organized by `_common_args.py`, `_tasks_args.py`, `_agents_args.py`, `_simple_args.py`, `_workflows_org_args.py`, `_remaining_args.py`).
+**Typed args (#1611 Phase 4)**: every domain registration in `src/synthorg/meta/mcp/domains/<domain>.py` now passes `args_model=<DomainArgs>` to the `read_tool` / `write_tool` / `admin_tool` factories. The `MCPToolInvoker` validates the raw arguments dict against the Pydantic model **before** dispatching to the handler; a malformed shape returns a typed `invalid_argument` envelope without ever invoking the handler. Migrated handlers therefore do NOT need to call `require_arg` / `require_non_blank` / `coerce_pagination` for fields the model already validates. Args models live in domain-scoped `_*_args.py` modules under `src/synthorg/meta/mcp/domains/`, with shared mixins (`PaginationFields`, `DestructiveGuardrailFields`) in `_common_args.py`. `DomainToolDef` runs a `_validate_args_model_property_alignment` model_validator post-construction that compares `args_model.model_fields` against the wire schema's `parameters["properties"]`; drift between invoker validation and the client-visible schema raises `ValueError` at registration time.
 
 Handlers run `require_destructive_guardrails(arguments, actor)` on any `admin_tool`, route through service-layer facades (never into `app_state.persistence.*` directly), and emit the three log paths via `common_logging` helpers (no per-handler `_log_*` duplicates).
