+++ src/synthorg/meta/mcp/domains/_remaining_args/__init__.py
@@ -1,0 +1,227 @@
+"""Typed argument models for the remaining MCP domains.
+
+Covers ``communication`` / ``integrations`` / ``infrastructure`` /
+``memory`` (the largest set, ~95 tools combined).  Models are split
+across four submodules to keep each file under the 800-line guideline:
+
+* :mod:`._communication` -- messages, meetings, connections, webhooks,
+  tunnel.
+* :mod:`._integrations` -- mcp_catalog, oauth, clients, artifacts,
+  ontology.
+* :mod:`._infrastructure` -- health, settings, providers, backup,
+  audit, events, users, projects, requests, setup, simulations,
+  template_packs, integration_health.
+* :mod:`._memory_finetune` -- fine-tuning, checkpoints, embedder,
+  GDPR delete.
+
+Re-exports keep the historical
+``from synthorg.meta.mcp.domains._remaining_args import X`` import
+surface stable for the four domain consumers
+(``communication.py`` / ``integrations.py`` / ``infrastructure.py`` /
+``memory.py``) and the smoke test in
+``tests/unit/meta/mcp/domains/test_args_smoke.py``.
+"""
+
+from synthorg.meta.mcp.domains._remaining_args._communication import (
+    ConnectionsCheckHealthArgs,
+    ConnectionsCreateArgs,
+    ConnectionsDeleteArgs,
+    ConnectionsGetArgs,
+    ConnectionsListArgs,
+    MeetingsCreateArgs,
+    MeetingsDeleteArgs,
+    MeetingsGetArgs,
+    MeetingsListArgs,
+    MeetingsUpdateArgs,
+    MessagesDeleteArgs,
+    MessagesGetArgs,
+    MessagesListArgs,
+    MessagesSendArgs,
+    TunnelConnectArgs,
+    TunnelGetStatusArgs,
+    WebhooksCreateArgs,
+    WebhooksDeleteArgs,
+    WebhooksGetArgs,
+    WebhooksListArgs,
+    WebhooksUpdateArgs,
+)
+from synthorg.meta.mcp.domains._remaining_args._infrastructure import (
+    AuditListArgs,
+    BackupCreateArgs,
+    BackupDeleteArgs,
+    BackupGetArgs,
+    BackupListArgs,
+    BackupRestoreArgs,
+    EventsListArgs,
+    HealthCheckArgs,
+    IntegrationHealthGetAllArgs,
+    IntegrationHealthGetArgs,
+    ProjectsCreateArgs,
+    ProjectsDeleteArgs,
+    ProjectsGetArgs,
+    ProjectsListArgs,
+    ProjectsUpdateArgs,
+    ProvidersGetArgs,
+    ProvidersGetHealthArgs,
+    ProvidersListArgs,
+    ProvidersTestConnectionArgs,
+    RequestsCreateArgs,
+    RequestsGetArgs,
+    RequestsListArgs,
+    SettingsDeleteArgs,
+    SettingsGetArgs,
+    SettingsListArgs,
+    SettingsUpdateArgs,
+    SetupGetStatusArgs,
+    SetupInitializeArgs,
+    SimulationsCreateArgs,
+    SimulationsGetArgs,
+    SimulationsListArgs,
+    TemplatePacksGetArgs,
+    TemplatePacksInstallArgs,
+    TemplatePacksListArgs,
+    TemplatePacksUninstallArgs,
+    UsersCreateArgs,
+    UsersDeleteArgs,
+    UsersGetArgs,
+    UsersListArgs,
+    UsersUpdateArgs,
+)
+from synthorg.meta.mcp.domains._remaining_args._integrations import (
+    ArtifactsCreateArgs,
+    ArtifactsDeleteArgs,
+    ArtifactsGetArgs,
+    ArtifactsListArgs,
+    ClientsCreateArgs,
+    ClientsDeactivateArgs,
+    ClientsGetArgs,
+    ClientsGetSatisfactionArgs,
+    ClientsListArgs,
+    McpCatalogGetArgs,
+    McpCatalogInstallArgs,
+    McpCatalogListArgs,
+    McpCatalogSearchArgs,
+    McpCatalogUninstallArgs,
+    OauthConfigureProviderArgs,
+    OauthListProvidersArgs,
+    OauthRemoveProviderArgs,
+    OntologyGetEntityArgs,
+    OntologyGetRelationshipsArgs,
+    OntologyListEntitiesArgs,
+    OntologySearchArgs,
+)
+from synthorg.meta.mcp.domains._remaining_args._memory_finetune import (
+    FineTuneBackend,
+    FineTuneExecutionConfig,
+    MemoryCancelFineTuneArgs,
+    MemoryDeleteCheckpointArgs,
+    MemoryDeleteEntryArgs,
+    MemoryDeployCheckpointArgs,
+    MemoryGetActiveEmbedderArgs,
+    MemoryGetFineTuneStatusArgs,
+    MemoryListCheckpointsArgs,
+    MemoryListRunsArgs,
+    MemoryResumeFineTuneArgs,
+    MemoryRollbackCheckpointArgs,
+    MemoryRunPreflightArgs,
+    MemoryStartFineTuneArgs,
+)
+
+__all__ = [
+    "ArtifactsCreateArgs",
+    "ArtifactsDeleteArgs",
+    "ArtifactsGetArgs",
+    "ArtifactsListArgs",
+    "AuditListArgs",
+    "BackupCreateArgs",
+    "BackupDeleteArgs",
+    "BackupGetArgs",
+    "BackupListArgs",
+    "BackupRestoreArgs",
+    "ClientsCreateArgs",
+    "ClientsDeactivateArgs",
+    "ClientsGetArgs",
+    "ClientsGetSatisfactionArgs",
+    "ClientsListArgs",
+    "ConnectionsCheckHealthArgs",
+    "ConnectionsCreateArgs",
+    "ConnectionsDeleteArgs",
+    "ConnectionsGetArgs",
+    "ConnectionsListArgs",
+    "EventsListArgs",
+    "FineTuneBackend",
+    "FineTuneExecutionConfig",
+    "HealthCheckArgs",
+    "IntegrationHealthGetAllArgs",
+    "IntegrationHealthGetArgs",
+    "McpCatalogGetArgs",
+    "McpCatalogInstallArgs",
+    "McpCatalogListArgs",
+    "McpCatalogSearchArgs",
+    "McpCatalogUninstallArgs",
+    "MeetingsCreateArgs",
+    "MeetingsDeleteArgs",
+    "MeetingsGetArgs",
+    "MeetingsListArgs",
+    "MeetingsUpdateArgs",
+    "MemoryCancelFineTuneArgs",
+    "MemoryDeleteCheckpointArgs",
+    "MemoryDeleteEntryArgs",
+    "MemoryDeployCheckpointArgs",
+    "MemoryGetActiveEmbedderArgs",
+    "MemoryGetFineTuneStatusArgs",
+    "MemoryListCheckpointsArgs",
+    "MemoryListRunsArgs",
+    "MemoryResumeFineTuneArgs",
+    "MemoryRollbackCheckpointArgs",
+    "MemoryRunPreflightArgs",
+    "MemoryStartFineTuneArgs",
+    "MessagesDeleteArgs",
+    "MessagesGetArgs",
+    "MessagesListArgs",
+    "MessagesSendArgs",
+    "OauthConfigureProviderArgs",
+    "OauthListProvidersArgs",
+    "OauthRemoveProviderArgs",
+    "OntologyGetEntityArgs",
+    "OntologyGetRelationshipsArgs",
+    "OntologyListEntitiesArgs",
+    "OntologySearchArgs",
+    "ProjectsCreateArgs",
+    "ProjectsDeleteArgs",
+    "ProjectsGetArgs",
+    "ProjectsListArgs",
+    "ProjectsUpdateArgs",
+    "ProvidersGetArgs",
+    "ProvidersGetHealthArgs",
+    "ProvidersListArgs",
+    "ProvidersTestConnectionArgs",
+    "RequestsCreateArgs",
+    "RequestsGetArgs",
+    "RequestsListArgs",
+    "SettingsDeleteArgs",
+    "SettingsGetArgs",
+    "SettingsListArgs",
+    "SettingsUpdateArgs",
+    "SetupGetStatusArgs",
+    "SetupInitializeArgs",
+    "SimulationsCreateArgs",
+    "SimulationsGetArgs",
+    "SimulationsListArgs",
+    "TemplatePacksGetArgs",
+    "TemplatePacksInstallArgs",
+    "TemplatePacksListArgs",
+    "TemplatePacksUninstallArgs",
+    "TunnelConnectArgs",
+    "TunnelGetStatusArgs",
+    "UsersCreateArgs",
+    "UsersDeleteArgs",
+    "UsersGetArgs",
+    "UsersListArgs",
+    "UsersUpdateArgs",
+    "WebhooksCreateArgs",
+    "WebhooksDeleteArgs",
+    "WebhooksGetArgs",
+    "WebhooksListArgs",
+    "WebhooksUpdateArgs",
+]
